Default Interface
In Java earlier versions, before 1.8, it was not possible to create any method with a body in Interface, 
the only abstract method with public access modifier was supported, but now from version 1.8, it is 
possible to declare default method that has body also. 
If the programmer wants to add new service or logic in the interface without doing modification on 
the implementation class of interface, then now from Java version 1.8 he/she can use default methods 
in the interface.